requirements.txt
blitz-bayesian-pytorch==0.2.7
grad-cam==1.3.7
h5py==2.10.0
imageio==2.9.0
matplotlib==3.5.1
nibabel==3.2.1
nilearn==0.7.1
nipype==1.6.0
numpy==1.19.5
opencv-contrib-python-headless==4.5.3.56
opencv-python==4.5.5.64
pandas==1.2.3
Pillow==8.2.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
SimpleITK==1.2.4
simplejson==3.17.6
sklearn==0.0
tensorboard==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.3.17
timm==0.4.12
torchio==0.18.76
torchsummary==1.5.1
tqdm==4.59.0
virtualenv==20.4.3